renamed project, added assertion
