+++ src/main/kotlin/org/ironica/amatsukaze/PlaygroundVisitor.kt
@@ -1,13 +1,12 @@
-package org.ironica.playground
+package org.ironica.amatsukaze
 
 import org.antlr.v4.runtime.tree.*
 import playgroundGrammarParser
 
-import org.ironica.playground.SpecialRetVal.*
-import org.ironica.playground.Type.*
-import org.ironica.playground.Variability.*
+import org.ironica.amatsukaze.SpecialRetVal.*
+import org.ironica.amatsukaze.Type.*
+import org.ironica.amatsukaze.Variability.*
 import playgroundGrammarVisitor
-import kotlin.math.exp
 import kotlin.math.pow
 
 enum class SpecialRetVal {
