+++ src/main/resources/application.conf
@@ -4,6 +4,6 @@ ktor {
         port = ${?PORT}
     }
     application {
-        modules = [ org.ironica.playground.ApplicationKt.module ]
+        modules = [org.ironica.amatsukaze.ApplicationKt.module]
     }
 }
